Migrate Utilities class into the CLI tool.
