+++ djvm/cli/src/main/kotlin/net/corda/djvm/tools/cli/TreeCommand.kt
@@ -1,6 +1,5 @@
 package net.corda.djvm.tools.cli
 
-import net.corda.djvm.tools.Utilities.workingDirectory
 import picocli.CommandLine.Command
 import java.nio.file.Files
